added new file for article route
