+++ backend/src/server.js
@@ -1,12 +1,12 @@
 import express from 'express'
-import router from './routes/route.js'
-import userRoute from './routes/userRoute.js'
-import 'dotenv/config'
 import mongoConnect from './db/db.js'
+import 'dotenv/config'
+import userRoute from './routes/userRoute.js'
+import articleRoute from './routes/articleRoute.js'
 
 const app = express()
 app.use(express.json())
-app.use(router)
+app.use(articleRoute)
 app.use(userRoute)
 
 const PORT = process.env.PORT || 4000
